requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-jsonfield==1.4.1
django-mysql==3.11.1
djangorestframework==3.12.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.3.4
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyparsing==2.4.6
python-dotenv==0.17.1
pytoml==0.1.21
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
webencodings==0.5.1
